1
These are the translated files for Pillars of Eternity. They were translated by running the en text files through a code built on replit using OpenAI api calls. this process took three days of non-stop work. I have no idea how to code so it's still pretty amazing what we are able to do with AI now. The API calls cost a total of 18$. NEED HELP, with modding Pillars of Eternity so that it can actually read these files and output them in Farsi lettering. Without access to a Persian font the game just displays a bunch of GUI errors.